Create custom plane for Volume Visual
