+++ src/VolumeVisual.js
@@ -2,10 +2,8 @@ import * as THREE from 'three';
 import utils from './utils';
 import VolumeMesh from './gfx/VolumeMesh';
 import VolumeBounds from './gfx/VolumeBounds';
+import VolumeFarPlane from './gfx/VolumeFarPlane';
 import Visual from './Visual';
-import VolumeMaterial from './gfx/shaders/VolumeMaterial';
-import gfxutils from './gfx/gfxutils';
-import meshes from './gfx/meshes/meshes';
 import settings from './settings';
 
 function VolumeVisual(name, dataSource) {
@@ -19,57 +17,11 @@ function VolumeVisual(name, dataSource) {
   this.add(this._frame.getMesh());
   this.showFrame(settings.now.modes.VD.frame);
 
-  this.buildFarPlane();
+  this._farPlane = new VolumeFarPlane(this._mesh, 2, 2);
+  this.add(this._farPlane.getMesh());
 }
 
 utils.deriveClass(VolumeVisual, Visual);
-
-// Thes geometric far plane is required for correct filling in the BFTexture in case, when far plane cuts the volume
-// cube. In cut place of cube there is no correct data in BFTexture and volume rendering integral is calculated
-// with errors.
-// Far plane cuts the cube in case of large volume scale (zoom), because farplane doesn't change
-VolumeVisual.prototype.buildFarPlane = function () {
-  // create plane with unit corners coords (for future rescale in vshader according to camera properties)
-  const planeGeo = new THREE.PlaneGeometry(2, 2, 1, 1);
-  const mat = VolumeMaterial.BackFacePosMaterialFarPlane();
-  this._plane = new meshes.Mesh(planeGeo, mat);
-  this._plane.frustumCulled = false;
-  this._plane.doubleSided = true;
-  const matWorldToVolume = new THREE.Matrix4();
-
-  this._plane._onBeforeRender = function (_renderer, _scene, camera, _geometry, _material, _group) {
-    const volume = this.parent.getObjectByName('VolumeMesh');
-    const { material } = this;
-    if (!volume || !material) {
-      return;
-    }
-
-    // count point in world at farplane place
-    const planeCamPos = new THREE.Vector4(0, 0, -(camera.far - 0.1), 1);
-    planeCamPos.applyMatrix4(camera.matrixWorld);
-
-    // recalc matrices to make plane be placed as farplane in the World relative to camera
-    this.matrix.identity();
-    this.matrix.makeTranslation(planeCamPos.x, planeCamPos.y, planeCamPos.z);
-    this.matrixWorld.copy(this.matrix);
-    this.modelViewMatrix.multiplyMatrices(camera.matrixWorldInverse, this.matrixWorld);
-    this.normalMatrix.getNormalMatrix(this.modelViewMatrix);
-
-    // build worldToVolume matrix to transform plane into volumeCS (volumeCS coords are written to BackFaceTexture)
-    const volumeMatrix = this.parent.getObjectByName('VolumeMesh').matrixWorld;
-    matWorldToVolume.getInverse(volumeMatrix);
-
-    // update material props
-    material.uniforms.aspectRatio.value = camera.aspect;
-    material.uniforms.farZ.value = camera.far;
-    material.uniforms.tanHalfFOV.value = Math.tan(THREE.Math.DEG2RAD * 0.5 * camera.fov);
-    material.uniforms.matWorld2Volume.value = matWorldToVolume;
-  };
-  this.add(this._plane);
-
-  // set it to special layer to draw only into BFTexture
-  this._plane.layers.set(gfxutils.LAYERS.VOLUME_BFPLANE);
-};
 
 VolumeVisual.prototype.getBoundaries = function () {
   const box = this._dataSource.getBox();
